+++ dev-site-config/site.config.js
@@ -5,6 +5,7 @@ const siteConfig = {
     themes: {
       'Default Theme': '',
       'Clinical Low Light Theme': 'clinical-lowlight-theme',
+      'Orion Fusion Theme': 'orion-fusion-theme',
     },
   },
   navConfig,