+++ packages/terra-tabs/tests/wdio/tabs-spec.js
@@ -1,5 +1,4 @@
 // Verify tabs collapse appropriately
-
 Terra.describeViewports('Tabs - Responsive', ['tiny', 'small', 'medium', 'large', 'huge', 'enormous'], () => {
   describe('Default', () => {
     it('Default', () => {
@@ -19,7 +18,7 @@ Terra.describeViewports('Tabs - Responsive', ['tiny', 'small', 'medium', 'large'
   });
   describe('Icon Only Tabs', () => {
     it('Icon Only Tabs', () => {
-      browser.url('/#/raw/tests/terra-tabs/tabs/tabs/icon-only-tabs');
+      browser.url('/#/raw/tests/terra-tabs/tabs/tabs/default-icon-only-tabs');
       browser.moveToObject('[class*="tab-content"]');
       Terra.validates.element();
     });
@@ -75,6 +74,45 @@ Terra.describeViewports('Tabs - Large screen', ['large'], () => {
       browser.url('/#/raw/tests/terra-tabs/tabs/tabs/fill-parent-tabs');
       browser.moveToObject('[class*="tab-content"]');
       Terra.validates.element();
+    });
+  });
+});
+
+// Verify tabs do not _completely_ collapse.
+Terra.describeViewports('Tabs - Uncollapsed', ['tiny'], () => {
+  describe('Icon Only Tabs', () => {
+    it('Icon Only Tabs', () => {
+      browser.url('/#/raw/tests/terra-tabs/tabs/tabs/uncollapsed-icon-only-tabs');
+      Terra.validates.element('uncollapsed', { selector: '#iconOnlyTabs' });
+    });
+  });
+
+  describe('Default', () => {
+    it('Default', () => {
+      browser.url('/#/raw/tests/terra-tabs/tabs/tabs/uncollapsed-tabs');
+      Terra.validates.element('uncollapsed');
+    });
+  });
+});
+
+Terra.describeViewports('Tabs - Responsive to Window', ['tiny', 'small', 'medium', 'large', 'huge', 'enormous'], () => {
+  describe('Responsive to Window', () => {
+    it('Responsive to Window', () => {
+      browser.url('/#/raw/tests/terra-tabs/tabs/tabs/tabs-responsive-to-window');
+      browser.moveToObject('[class*="tab-content"]');
+      Terra.validates.element({ selector: '#tabs-container' });
+    });
+  });
+});
+
+Terra.describeViewports('Tabs - Responsive to Parent', ['huge'], () => {
+  describe('Responsive to Parent', () => {
+    it('Responsive to Parent', () => {
+      browser.url('/#/raw/tests/terra-tabs/tabs/tabs/tabs-responsive-to-parent');
+      Terra.validates.element('before');
+      browser.click('button');
+      browser.moveToObject('[class*="tab-content"]');
+      Terra.validates.element('after');
     });
   });
 });