+++ packages/terra-slide-panel-manager/CHANGELOG.md
@@ -5,6 +5,7 @@ Unreleased
 ----------
 ### Added
 * Added reference wdio screenshots for "clinical-lowlight-theme".
+* Added reference wdio screenshots for "orion-fusion-theme".
 
 5.19.0 - (November 20, 2019)
 ------------------